Changes:
Make class to an object.
The idea is to have something similar to Ruby
but without having the Class public to the user.

Add method to add user-defined methods.

Add the allocator func.
This function will be defined in the object.go file
each class that has its own allocator need to
define it.