Change: Check member selector
Rename FieldSel to MemberSelector
Member seletor represets an access to a object's member
it can be a field or method

Rename CallExpr node to MethodCallExpr
It has a MemberSelector -> represents the "receiver.member"
and it also has a slice of expr for Arguments

Add new node named FunctionCallExpr
it has the function name and a slice of expr for Arguments.
This node will be used to implicity call on the current object (a.k.a
this)

Typecheck MemberSelector when accessing a Field and Method in a
MethodCallExpr.