make each parameter of a function a VarDecl.
I am not so sure about this change.
I thought of having a struct `NameType` and using it.
With the NameType struct, I could change the VarDecl node to:

```
type VarDecl struct {
    *NameType
    Value Expr
}
```

I didn't use NameType because I want to allow `default arguments` and
maybe VarDecl is the way to go since parameters are implict variable
declarations.